Update to nightly-2025-11-02
A bunch of error messages changed.

This also brings in thread IDs in panic messages, which we need to normalize
away. I first tried implementing this normalization without regex, but
it didn't work first try, so I chose regex, which is simpler.

Also updated the builder to edition 2024 (since my initial
implementation required let chains) which requires using unsafe for
`remove_var`, but that's trivial unsafe that's fine.

I also added a note about the current version to the index.